Update Android app configuration and API service
This commit updates the NDK version in the build.gradle.kts file to a specific version. Additionally, it modifies the server host URL in config.dart to a new ngrok address, ensuring the app connects to the correct server. The api_service.dart file is also updated to remove the unused raspberryPiBaseUrl variable, streamlining the code.